refactor: modularize finalization step
